Beta Release v0.2.28

+++ managed/src/SwiftlyS2.Core/Modules/GameData/GameDataService.cs
@@ -104,7 +104,7 @@ internal class GameDataService : IGameDataService {
     return NativeOffsets.Exists(offsetName);
   }
 
-  public nint GetOffset(string offsetName) {
+  public int GetOffset(string offsetName) {
     if (_Offsets.TryGetValue(offsetName, out var offset)) {
       return offset;
     }
